feat: add browser build without ssr code
